refactor: 결과 판단 로직 Players로 이동
